+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/Governor/KnownPeers.hs
@@ -9,7 +9,6 @@ module Ouroboros.Network.PeerSelection.Governor.KnownPeers
   , aboveTarget
   ) where
 
-import qualified Data.Map as Map
 import           Data.Maybe (fromMaybe)
 import qualified Data.Set as Set
 
@@ -21,6 +20,8 @@ import           Control.Monad.Class.MonadTime.SI
 import           Control.Monad.Class.MonadTimer.SI
 
 import           Ouroboros.Network.PeerSelection.Governor.Types
+import           Ouroboros.Network.PeerSelection.PeerAdvertise
+                     (PeerAdvertise (..))
 import           Ouroboros.Network.PeerSelection.PeerSharing (PeerSharing (..))
 import qualified Ouroboros.Network.PeerSelection.State.EstablishedPeers as EstablishedPeers
 import           Ouroboros.Network.PeerSelection.State.KnownPeers
@@ -187,15 +188,19 @@ jobPeerShare PeerSelectionActions{requestPeerShare}
                                           ]
                         , decisionState =
                            st { -- TODO: also update with the failures
-                                knownPeers = KnownPeers.insert
-                                               (Map.fromList
-                                                $ map (\a -> ( a
-                                                             , ( Nothing
-                                                               , Nothing
-                                                               , Nothing))
-                                                      )
-                                                      newPeers)
-                                               (knownPeers st),
+                                knownPeers = KnownPeers.alter
+                                              (\x -> case x of
+                                                Nothing ->
+                                                  KnownPeers.alterKnownPeerInfo
+                                                    (Nothing, Just DoAdvertisePeer, Nothing)
+                                                    x
+                                                Just _ ->
+                                                  KnownPeers.alterKnownPeerInfo
+                                                    (Nothing, Nothing, Nothing)
+                                                    x
+                                              )
+                                              (Set.fromList newPeers)
+                                              (knownPeers st),
                                 inProgressPeerShareReqs = inProgressPeerShareReqs st
                                                         - length peers
                            }
@@ -231,15 +236,19 @@ jobPeerShare PeerSelectionActions{requestPeerShare}
                                            ]
                          , decisionState =
                             st { -- TODO: also update with the failures
-                                 knownPeers = KnownPeers.insert
-                                                (Map.fromList
-                                                 $ map (\a -> ( a
-                                                              , ( Nothing
-                                                                , Nothing
-                                                                , Nothing))
-                                                       )
-                                                       newPeers)
-                                                (knownPeers st),
+                                 knownPeers = KnownPeers.alter
+                                               (\x -> case x of
+                                                 Nothing ->
+                                                   KnownPeers.alterKnownPeerInfo
+                                                     (Nothing, Just DoAdvertisePeer, Nothing)
+                                                     x
+                                                 Just _ ->
+                                                   KnownPeers.alterKnownPeerInfo
+                                                     (Nothing, Nothing, Nothing)
+                                                     x
+                                               )
+                                               (Set.fromList newPeers)
+                                               (knownPeers st),
                                  inProgressPeerShareReqs = inProgressPeerShareReqs st
                                                          - length peerResults
                                }
@@ -289,21 +298,24 @@ jobPeerShare PeerSelectionActions{requestPeerShare}
                 Left partialResults -> [ p | Just (Right (PeerSharingResult ps)) <- partialResults
                                            , p <- ps
                                            , not (isKnownLedgerPeer p (knownPeers st)) ]
-
          in Decision { decisionTrace = [ TracePeerShareResults peerResults
                                        , TracePeerShareResultsFiltered newPeers
                                        ]
                      , decisionState =
                         st { -- TODO: also update with the failures
-                             knownPeers = KnownPeers.insert
-                                            (Map.fromList
-                                             $ map (\a -> ( a
-                                                          , ( Nothing
-                                                            , Nothing
-                                                            , Nothing))
-                                                   )
-                                                   newPeers)
-                                            (knownPeers st),
+                             knownPeers = KnownPeers.alter
+                                           (\x -> case x of
+                                             Nothing ->
+                                               KnownPeers.alterKnownPeerInfo
+                                                 (Nothing, Just DoAdvertisePeer, Nothing)
+                                                 x
+                                             Just _ ->
+                                               KnownPeers.alterKnownPeerInfo
+                                                 (Nothing, Nothing, Nothing)
+                                                 x
+                                           )
+                                           (Set.fromList newPeers)
+                                           (knownPeers st),
                              inProgressPeerShareReqs = inProgressPeerShareReqs st
                                                      - length peers
                            }