+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/State/KnownPeers.hs
@@ -7,10 +7,13 @@ module Ouroboros.Network.PeerSelection.State.KnownPeers
   ( -- * Types
     KnownPeers
   , invariant
+    -- * KnownPeerInfo operations
+  , alterKnownPeerInfo
     -- * Basic container operations
   , empty
   , size
   , insert
+  , alter
   , delete
   , toSet
   , member
@@ -22,6 +25,7 @@ module Ouroboros.Network.PeerSelection.State.KnownPeers
   , lookupTepidFlag
   , setTepidFlag
   , clearTepidFlag
+  , setSuccessfulConnectionFlag
     -- ** Tracking when we can (re)connect
   , minConnectTime
   , setConnectTimes
@@ -93,7 +97,7 @@ data KnownPeerInfo = KnownPeerInfo {
        -- It is used to implement the exponential backoff strategy and may also
        -- be used by policies to select peers to forget.
        --
-       knownPeerFailCount :: !Int,
+       knownPeerFailCount        :: !Int,
 
        -- | Indicates if the peer was hot but then got demoted.
        --
@@ -104,7 +108,7 @@ data KnownPeerInfo = KnownPeerInfo {
        -- It is also used as useful information for the Peer Selection Governor
        -- when deciding which peers to share when Peer Sharing.
        --
-       knownPeerTepid     :: !Bool,
+       knownPeerTepid            :: !Bool,
 
        -- | Indicates current remote Peer Willingness information.
        --
@@ -113,7 +117,7 @@ data KnownPeerInfo = KnownPeerInfo {
        --
        -- It is used by the Peer Sharing logic to decide if we should share/ask
        -- about/to this peer's address to others.
-       knownPeerSharing   :: !PeerSharing,
+       knownPeerSharing          :: !PeerSharing,
 
        -- | Indicates current local Peer Willingness information.
        --
@@ -122,7 +126,7 @@ data KnownPeerInfo = KnownPeerInfo {
        --
        -- It is used by the Peer Sharing logic to decide if we should share
        -- about this peer's address to others.
-       knownPeerAdvertise :: !PeerAdvertise,
+       knownPeerAdvertise        :: !PeerAdvertise,
 
        -- | Indicates if peer came from ledger.
        --
@@ -130,7 +134,16 @@ data KnownPeerInfo = KnownPeerInfo {
        -- reply, since ledger peers are not particularly what one is looking for
        -- in a Peer Sharing reply.
        --
-       knownLedgerPeer    :: !IsLedgerPeer
+       knownLedgerPeer           :: !IsLedgerPeer,
+
+       -- | Indicates if the node managed to connect to the peer at some point
+       -- in time.
+       --
+       -- This differs from the tepid flag in a way that this flag will be
+       -- set/enabled if we established a successful connection with this
+       -- peer. It won't be unset after this.
+       --
+       knownSuccessfulConnection :: !Bool
      }
   deriving (Eq, Show)
 
@@ -148,6 +161,32 @@ invariant KnownPeers{..} =
          availableToConnect
         (Set.fromList (PSQ.keys nextConnectTimes)))
 
+
+-------------------------------
+-- KnownPeerInfo manipulation
+--
+
+alterKnownPeerInfo
+  :: (Maybe PeerSharing, Maybe PeerAdvertise, Maybe IsLedgerPeer)
+  -> Maybe KnownPeerInfo
+  -> Maybe KnownPeerInfo
+alterKnownPeerInfo (peerSharing, peerAdvertise, ledgerPeers) peerLookupResult =
+  case peerLookupResult of
+    Nothing -> Just $
+      KnownPeerInfo {
+        knownPeerFailCount = 0
+      , knownPeerTepid     = False
+      , knownPeerSharing   = fromMaybe PeerSharingDisabled peerSharing
+      , knownPeerAdvertise = fromMaybe DoNotAdvertisePeer peerAdvertise
+      , knownLedgerPeer    = fromMaybe IsNotLedgerPeer ledgerPeers
+      , knownSuccessfulConnection = False
+      }
+    Just kpi -> Just $
+      kpi {
+        knownPeerSharing   = fromMaybe (knownPeerSharing kpi) peerSharing
+      , knownPeerAdvertise = fromMaybe (knownPeerAdvertise kpi) peerAdvertise
+      , knownLedgerPeer    = fromMaybe (knownLedgerPeer kpi) ledgerPeers
+      }
 
 -------------------------------
 -- Basic container operations
@@ -175,7 +214,17 @@ member :: Ord peeraddr
 member peeraddr KnownPeers {allPeers} =
     peeraddr `Map.member` allPeers
 
--- TODO: `insert` ought to be idempotent, see issue #4616.
+-- | This inserts a map of peers with its respective peer sharing, peer
+-- advertise and ledger flags into the known peers set.
+--
+-- Please note that if in the map there's an entry for a peer already present
+-- in the known peers set, then its values will only be overwritten if they
+-- are a 'Just'. Otherwise the current information will be preserved. On the
+-- other hand if there's an entry for a peer that isn't a member of the known
+-- peer set, the 'Nothing' values will default to 'NoPeerSharing',
+-- 'DoNotAdvertisePeer' and 'IsNotLedgerPeer', respectively, unless a 'Just'
+-- value is used.
+--
 insert :: Ord peeraddr
        => Map peeraddr (Maybe PeerSharing, Maybe PeerAdvertise, Maybe IsLedgerPeer)
        -> KnownPeers peeraddr
@@ -187,39 +236,41 @@ insert peeraddrs
        } =
     let allPeersAddrs = Map.keysSet peeraddrs
         knownPeers' = knownPeers {
-          allPeers =
-              let (<+>) = Map.unionWith mergePeerInfo in
-              allPeers
-          <+> Map.map newPeerInfo peeraddrs,
-
+          allPeers = Map.foldlWithKey' (\m peer v -> Map.alter (alterKnownPeerInfo v) peer m)
+                                       allPeers
+                                       peeraddrs,
           availableToConnect =
               availableToConnect
            <> Set.filter (`Map.notMember` allPeers) allPeersAddrs
         }
     in assert (invariant knownPeers') knownPeers'
-  where
-    newPeerInfo (peerSharing, peerAdvertise, ledgerPeers) =
-      let peerAdvertise' = fromMaybe DoNotAdvertisePeer peerAdvertise
-          peerSharing'   = fromMaybe PeerSharingDisabled peerSharing
-       in KnownPeerInfo {
-        knownPeerFailCount = 0
-      , knownPeerTepid     = False
-      , knownPeerSharing   = peerSharing'
-      , knownPeerAdvertise = peerAdvertise'
-      , knownLedgerPeer    = fromMaybe IsNotLedgerPeer ledgerPeers
-      }
-    mergePeerInfo old new =
-      KnownPeerInfo {
-        knownPeerFailCount = knownPeerFailCount old
-      , knownPeerTepid     = knownPeerTepid old
-      -- It might be the case we are updating a peer's particular willingness
-      -- flags or we just learned this peer comes from ledger.
-      , knownPeerSharing   = knownPeerSharing new
-      , knownPeerAdvertise = knownPeerAdvertise new
-      -- Preserve Ledger Peer information if the peer is ledger.
-      , knownLedgerPeer    = case knownLedgerPeer old of
-                               IsLedgerPeer    -> IsLedgerPeer
-                               IsNotLedgerPeer -> knownLedgerPeer new
+
+alter :: Ord peeraddr
+      => (Maybe KnownPeerInfo -> Maybe KnownPeerInfo)
+      -> Set peeraddr
+      -> KnownPeers peeraddr
+      -> KnownPeers peeraddr
+alter f ks knownPeers@KnownPeers {
+            allPeers = allPeers
+          , availableToConnect = availableToConnect
+          , nextConnectTimes
+          } =
+  let newAllPeers =
+        Set.foldl' (\acc k -> Map.alter f k acc)
+                   allPeers
+                   ks
+      deletedPeers =
+        Set.filter (`Map.notMember` newAllPeers) ks
+      newAvailableToConnect =
+        (availableToConnect <> ks)
+        `Set.difference`
+        deletedPeers
+      newNextConnectTimes =
+        Set.foldl' (flip PSQ.delete) nextConnectTimes ks
+   in knownPeers {
+        allPeers           = newAllPeers
+      , availableToConnect = newAvailableToConnect
+      , nextConnectTimes   = newNextConnectTimes
       }
 
 delete :: Ord peeraddr
@@ -323,6 +374,16 @@ setTepidFlag' val peeraddr knownPeers@KnownPeers{allPeers} =
                               peeraddr allPeers
                }
 
+setSuccessfulConnectionFlag :: Ord peeraddr
+                             => peeraddr
+                             -> KnownPeers peeraddr
+                             -> KnownPeers peeraddr
+setSuccessfulConnectionFlag peeraddr knownPeers@KnownPeers{allPeers} =
+    assert (peeraddr `Map.member` allPeers) $
+    knownPeers { allPeers = Map.update (\kpi  -> Just kpi { knownSuccessfulConnection = True })
+                              peeraddr allPeers
+               }
+
 clearTepidFlag :: Ord peeraddr
              => peeraddr
              -> KnownPeers peeraddr
@@ -387,8 +448,10 @@ setConnectTimes times
 canPeerShareRequest :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> Bool
 canPeerShareRequest pa KnownPeers { allPeers } =
   case Map.lookup pa allPeers of
-    Just (KnownPeerInfo _ _ PeerSharingEnabled _ _) -> True
-    _                                               -> False
+    Just KnownPeerInfo
+          { knownPeerSharing = PeerSharingEnabled
+          } -> True
+    _     -> False
 
 -- Only share peers which are allowed to be advertised, i.e. have
 -- 'DoAdvertisePeer' 'PeerAdvertise' values.
@@ -396,8 +459,11 @@ canPeerShareRequest pa KnownPeers { allPeers } =
 canSharePeers :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> Bool
 canSharePeers pa KnownPeers { allPeers } =
   case Map.lookup pa allPeers of
-    Just (KnownPeerInfo _ _ _ DoAdvertisePeer _) -> True
-    _                                            -> False
+    Just KnownPeerInfo
+          { knownPeerAdvertise        = DoAdvertisePeer
+          , knownSuccessfulConnection = True
+          } -> True
+    _     -> False
 
 -- | Filter peers available for Peer Sharing requests, according to their
 -- 'PeerSharing' information
@@ -417,8 +483,11 @@ getPeerSharingResponsePeers :: KnownPeers peeraddr
 getPeerSharingResponsePeers knownPeers =
     Map.keysSet
   $ Map.filter (\case
-                  KnownPeerInfo _ _ _ DoAdvertisePeer _ -> True
-                  _                                     -> False
+                  KnownPeerInfo
+                    { knownPeerAdvertise        = DoAdvertisePeer
+                    , knownSuccessfulConnection = True
+                    } -> True
+                  _   -> False
                )
   $ allPeers knownPeers